Update to pages to take and check input
API functions now have a redirect to a success page so you know your request was processed (this will be expanded to include a fail page along with the error message that caused it and the recommended course of action (if the Quartzy server is down, we will need to direct the users to use paper for this request)).  Increase and decrease pages now have the ability to take user input and change the stock based on that input (only the quantity can be adjusted, I still need to find a way to store references between barcodes, item ID, and item name (I'm thinking a modifiable text file but I'm not sure yet)).  Additional user input for the database should be easy enough to implement and I will do so when I have the time.