Added some style
Text above input fields, align test left, input field center (same margin), etc.  Not a lot of changes, but it makes the app look a little bit better.  Buttons are currently wrapping around the entire container which I would like to change, but I'm having issues and it seems like the only solution is going to be putting the button in its own container.